zl38005: Oops bug fix
With some previous commits the memory allocation needed for reading
registers is 32bit.
Even if we need to read 16 bits it is mandatory to allocate an int variable,
because it is used by the lower levels to write 32 bits of data.

+++ include/sound/zl38005.h
@@ -16,6 +16,6 @@
 #define ZL38005_USRGAIN		0x046B
 
 int zl38005_write(u16 addr, u16 val);
-int zl38005_read(u16 addr, u16 *val);
+int zl38005_read(u16 addr, int *val);
 
 #endif
